+++ src/main/java/net/silentchaos512/gear/item/gear/CoreArmor.java
@@ -128,7 +128,7 @@ public class CoreArmor extends DyeableArmorItem implements ICoreArmor {
     public void setDamage(ItemStack stack, int damage) {
         if (GearHelper.isUnbreakable(stack))
             return;
-        if (!Config.GENERAL.gearBreaksPermanently.get())
+        if (!Config.Server.gearBreaksPermanently.get())
             damage = MathHelper.clamp(damage, 0, getMaxDamage(stack));
         super.setDamage(stack, damage);
     }
@@ -138,7 +138,7 @@ public class CoreArmor extends DyeableArmorItem implements ICoreArmor {
         int value;
         if (GearHelper.isUnbreakable(stack)) {
             value = 0;
-        } else if (!Config.GENERAL.gearBreaksPermanently.get()) {
+        } else if (!Config.Server.gearBreaksPermanently.get()) {
             value = MathHelper.clamp(amount, 0, stack.getMaxDamage() - stack.getDamage() - 1);
         } else {
             value = super.damageItem(stack, amount, entity, onBroken);