+++ src/main/java/net/silentchaos512/gear/util/GearData.java
@@ -130,7 +130,7 @@ public final class GearData {
                     TraitActionContext context = new TraitActionContext(player, level, stack);
                     return trait.onGetStat(context, stat, val, damageRatio);
                 });
-                final float value = Config.GENERAL.getStatWithMultiplier(stat, withTraits);
+                final float value = Config.Server.getStatWithMultiplier(stat, withTraits);
                 // SilentGear.log.debug(stat, value);
                 ResourceLocation statId = Objects.requireNonNull(stat.getRegistryName());
                 propertiesCompound.remove(statId.getPath()); // Remove old keys