+++ src/main/java/net/silentchaos512/gear/util/GearHelper.java
@@ -253,7 +253,7 @@ public final class GearHelper {
     }
 
     private static boolean canBreakPermanently(ItemStack stack) {
-        return Config.GENERAL.gearBreaksPermanently.get() || GearData.hasPart(stack, MiscUpgrades.RED_CARD.getPartId());
+        return Config.Server.gearBreaksPermanently.get() || GearData.hasPart(stack, MiscUpgrades.RED_CARD.getPartId());
     }
 
     public static boolean isBroken(ItemStack stack) {